+++ HISTORY.md
@@ -1,3 +1,8 @@
+## 1.1.3 (in progress)
+
+- CNV: support background inputs for CNVkit, GATK4 CNV and seq2c. Allows
+  pre-computed panel of normals for tumor-only or single sample CNV calling.
+
 ## 1.1.2 (12 December 2018)
 
 - VarDict low frequency somatic filters: generalize strand and mismatch based